requirements.txt
line-bot-sdk
flask    
bs4           
beautifulsoup4 == 4.1.7
imgurpython
